Release based pull request Encrypt all payloads to bypass AV detection #14
Encrypted all payloads with AES using the key "FREDDY". This key is stored in the extension settings and used to decrypt on initialization. This is done in an effort to prevent automatic decryption by AV's as the relationship between the stored key and loaded key is lost without the IHttpCallbacks implementation.

Signed-off-by: Steven van der Baan <steven@vdbaan.net>